Copy-Files改用robocopy (Copy-WithProgress) 因為Copy-Files資料名稱太長會無法複製

+++ Pad_backup/backup_pad_library.ps1
@@ -2,34 +2,36 @@
 Write-Host "Hello $Env:UserName"
 
 #region Check connection
-Test-PathExist $global:w_dir
+UtilityProgram\Test-PathExist $global:w_dir
 $box_dir = "C:\Users\$Env:UserName\Box"
-Test-PathExist $box_dir
+UtilityProgram\Test-PathExist $box_dir
 #endregion
 
 # backup destination
+$temp_backup_dir = "C:\Users\$Env:UserName\Documents\TempBackup\Backup_Flex_PAD\"
 $backup_dir = "$box_dir\Backup_Flex_PAD\"
-$dest_dir = UtilityProgram\Get-NowFolder $backup_dir
+$dest_dir = UtilityProgram\Get-NowFolder $temp_backup_dir
 
 #data source
-<# 批次複製 & 進度條
-$path_array = @( "Library-Intel", "Library-Nvida", "Library-Special", "Library-Flex")
-foreach ( $path in $path_array) {
-    Copy-WithProgress -Source "$w_dir\$path" -Destination "$dest_dir\$path"
-}
-#>
 
-$path_array = New-Object System.Collections.ArrayList
-foreach ( $path in @( "Library-Intel", "Library-Nvida", "Library-Special", "Library-Flex")) {
-    $path_array.Add("$global:w_dir\$path")
+# 批次複製 & 進度條
+$excluded_folders = @("footprint_building_aid_skill", "Library-Checking")
+$path_list = Get-ChildItem -Path $global:w_dir | 
+Where-Object { $_.Name -notin $excluded_folders } | 
+Select-Object -ExpandProperty FullName
+foreach ( $path in $path_list) {
+    UtilityProgram\Copy-WithPorgress -Source $path -Destination $dest_dir
 }
 
-UtilityProgram\Copy-Files -FilePath $path_array -Destination $dest_dir
-
-Compress-Folder $dest_dir
-
-# Delete oldest file when > 3 month
-Get-ChildItem -Path $backup_dir -Filter "*.zip" | Where-Object { ($_.LastWriteTime -lt (Get-Date).AddMonths(-3)) } | Remove-Item
+if ((UtilityProgram\Compress-Folder $dest_dir) -eq $true) {
+    $zip_file = "$dest_dir.zip"
+    # Copy to box
+    UtilityProgram\Copy-WithPorgress -Source $zip_file -Destination $backup_dir
+    # Delete temp file
+    Get-ChildItem -Path $zip_file | Remove-Item
+    # Delete oldest file when > 3 month
+    Get-ChildItem -Path $backup_dir -Filter "*.zip" | Where-Object { ($_.LastWriteTime -lt (Get-Date).AddMonths(-3)) } | Remove-Item
+}
 
 <#
 $dir = [System.Environment]::CurrentDirectory + "\Pad_backup"
